Create a first prototype of the interpreter
It took me some time to build it.

The internal state "receives" some actions (with associated data),
and it creates ad-hoc structures that implement the Operator
trait to modify itself.
I know, it's not the best thing ever, but the original design
(have operators that modify a given state) is not possible, as
trait objects must implement Sized, but then you cannot
use self, defeating the purpose of the design.